redesing of the app under mongodb + a major bugfix
now the app can communicate with mongodb, however only login and assinging levels works with it, the rest of the repository has to be yet rewritten under the new system.

A major error in the lessonrepostiry has also been fixed. Turned out that the whole command lacked functions managging different languages for the courses.